LayoutTests/platform/mac-mojave/fast/text/international/system-language/navigator-language/navigator-language-ru-expected.txt

PASS navigator.language is "ru-ru"
PASS successfullyParsed is true

TEST COMPLETE
